Improve.

+++ src/components/CartList.tsx
@@ -7,6 +7,7 @@ import { Button, Table } from 'react-bootstrap';
 import TotalAmount from '../components/ToalAmount';
 import CartProduct from '../components/CartProduct';
 import * as actions from '../actions';
+import Noty from 'noty';
 
 const CartList: React.FC<ProductListType & RouteComponentProps> = ({
   history,
@@ -14,14 +15,25 @@ const CartList: React.FC<ProductListType & RouteComponentProps> = ({
   const dispatch = useDispatch();
   const onCheckout = (id: number) => {
     dispatch(actions.Checkout({ id }));
-    history.push('/order');
+    history.push('/checkout');
   };
   const onCheckoutAll = () => {
-    dispatch(actions.Checkout({}));
-    history.push('/order');
+    const _products = products.filter((e) => e.chk);
+    if (_products.length > 0) {
+      dispatch(actions.Checkout({}));
+      history.push('/checkout');
+    } else {
+      new Noty({
+        type: 'error',
+        text: `선택한 장바구니 상품이 없습니다.`,
+        timeout: 3000,
+      }).show();
+    }
   };
   const onDeleteCart = (id: number) => {
-    dispatch(actions.DeleteCart({ id }));
+    if (window.confirm('정말 삭제하시겠습니까?')) {
+      dispatch(actions.DeleteCart({ id }));
+    }
   };
 
   const products = useSelector((state: RootState) => state.Shopping.cart);
@@ -52,7 +64,9 @@ const CartList: React.FC<ProductListType & RouteComponentProps> = ({
       </Table>
 
       <hr />
-      <TotalAmount products={products} />
+      <h4>
+        Total: <TotalAmount products={products} />
+      </h4>
 
       {products.length > 0 && (
         <Button size="lg" block onClick={onCheckoutAll}>
@@ -66,7 +80,7 @@ function NotFound() {
   return (
     <Fragment>
       <tr>
-        <td colSpan={5}>장바구니에 담겨진 상품이 없습니다.</td>
+        <td colSpan={6}>장바구니에 담겨진 상품이 없습니다.</td>
       </tr>
     </Fragment>
   );
